pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-chroma>=0.2.0",
    "langchain-community>=0.3.15",
    "langchain-openai>=0.3.1",
    "langgraph>=0.2.65",
]
